Unified navbars into a component
